Move some tests to printer subdirectory
