+++ test/unit/loops.spec.ts
@@ -595,30 +595,6 @@ test.each([
     expect(util.transpileString(loop, luajit).indexOf("::__continue1::") !== -1).toBe(true);
 });
 
-test("for dead code after return", () => {
-    const result = util.transpileAndExecute(`for (let i = 0; i < 10; i++) { return 3; const b = 8; }`);
-
-    expect(result).toBe(3);
-});
-
-test("for..in dead code after return", () => {
-    const result = util.transpileAndExecute(`for (let a in {"a": 5, "b": 8}) { return 3; const b = 8; }`);
-
-    expect(result).toBe(3);
-});
-
-test("for..of dead code after return", () => {
-    const result = util.transpileAndExecute(`for (let a of [1,2,4]) { return 3; const b = 8; }`);
-
-    expect(result).toBe(3);
-});
-
-test("while dead code after return", () => {
-    const result = util.transpileAndExecute(`while (true) { return 3; const b = 8; }`);
-
-    expect(result).toBe(3);
-});
-
 test("do...while", () => {
     const code = `
         let result = 0;
